Add files via upload
Änderungen:
- Neue Datei: crewliste.xlsx
- index_name_list: Namen sind nicht länger fest im Code sondern werden aus einer Excel-Datei ausgelesen. Dazu wird die Spalte „Namen“ genutzt. Wenn es in der Excel-Datei einen Spitznamen gibt, wird dieser anstatt des Namens verwendet. Dazu wir die Spalte „Spitznamen“ genutzt.
- person_capacity_list: Schichtanzahl der Personen sind nicht länger fest im Code sondern werden aus der Excel-Datei ausgelesen. Dazu wird die spalte „Schichtanzahl“ verwendet.  Wenn dort keine angegeben wird, wird automatisch 5 genommen. Dieser Wert „5“ ist noch fest im Code.
- preferred_shift_category_list: Präferenz ob Personen gerne beim Check_In arbeiten möchten ist nicht länger fest im Code sondern werden aus der Excel-Datei ausgelesen. Es werden die die Werte mit „CHECK_IN“ übergeben. Dazu weiterhin die Zuordnung funktioniert wird außerdem die Spalte „Nummer“ mitübergeben. Damit die Zuordnung auch mit den anderen Werten passt muss die Liste nach Vornamen alphabetisch sortiert sein. Die Werte in crewliste.xlsx sind als Tabelle formatiert, wodurch das einfach möglich ist.

changes:- New file: crewlist.xlsx- index_name_list: Names are no longer fixed in the code but are read from an Excel file. The column "Names" is used for this purpose. If there is a nickname in the Excel file, it is used instead of the name. The column "Nicknames" is used for this purpose.- person_capacity_list: The number of shifts of the persons are no longer fixed in the code but are read from the Excel file. The column "Number of shifts" is used for this purpose.  If none is specified there, 5 is automatically taken. This value "5" is still fixed in the code.- preferred_shift_category_list: Preference whether persons would like to work at Check_In is no longer fixed in the code but is read from the Excel file. The values are transferred with "CHECK_IN". To ensure that the assignment continues to work, the "Number" column is also transferred. The list must be sorted alphabetically by first name so that the assignment also fits with the other values. The values in crewliste.xlsx are formatted as a table, which makes this easy.Translated with www.DeepL.com/Translator (free version)